Introduce a tight type for internal _gather
IndexedFrame._gather and Index._gather are used in many places to
reorder or subsample from an existing frame or index. Sometimes, this
call to gather comes with some indices provided by the user, these
must be checked for in-boundsness, since libcudf does
not (deliberately) do so. Many other times, the arguments to _gather
come from other cudf functions that know they are constructing a
column that will be in-bounds as a gather map for the gatheree.

To date, cudf has taken a defensive approach in its implementation of
_gather. It accepts anything, attempts to coerce it to an integer
column, (optionally, but by default) checks that it is valid, and then
calls into libcudf.

It is easy to leave out the keyword argument to turn off the bounds
checking, and as a consequence there are a few places where we are
launching an unnecessary kernel to check the in-boundsness of
something that is a priori in bounds.

To address this, tighten the type of _gather to accept a GatherMap
object which encapsulates the information about bounds checking, and
push the responsibility for bounds checking onto the callers of
_gather.