Let gather map remember number of rows it is valid for
This way we can check if we're gathering something incorrectly. We
don't have existentially or universally qualified types, so we can't
statically construct a type that carries the proof that the gather map
is valid. As a cop-out, just remember the rows and check that the
thing we're being given to gather has an appropriate number of rows.